+++ tests/Models/UserAsArrayable.php
@@ -2,10 +2,8 @@
 
 namespace Tests\Models;
 
-use ArrayIterator;
-use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
-use Illuminate\Contracts\Support\Arrayable;
 use Staudenmeir\EloquentJsonRelations\Relations\BelongsToJson;
+use Tests\Casts\ArrayableCast;
 
 class UserAsArrayable extends User
 {
@@ -23,23 +21,5 @@ class UserAsArrayable extends User
     public function roles3(): BelongsToJson
     {
         return $this->belongsToJson(Role::class, 'options[]->role_id');
-    }
-}
-
-class ArrayableCast extends ArrayIterator implements CastsAttributes, Arrayable
-{
-    public function get($model, string $key, $value, array $attributes)
-    {
-        return null === $value ? new static([]) : new static(json_decode($value, true));
-    }
-
-    public function set($model, string $key, $value, array $attributes)
-    {
-        return $value instanceof static ? json_encode($value->toArray()) : json_encode($value);
-    }
-
-    public function toArray()
-    {
-        return $this->getArrayCopy();
     }
 }
